# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: FlowCraft API
  version: "1.0.0"
  description: FlowCraft platform API

# ---------- paths ----------
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        default:
          $ref: "#/components/responses/Error"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      responses:
        "204":
          description: Logged out
        default:
          $ref: "#/components/responses/Error"
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        default:
          $ref: "#/components/responses/Error"
...
components:
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
            code:
              type: string
              description: Machine-readable error code (platform + errdefs mapping).

    # ── Pagination ──
...
    LoginRequest:
      type: object
      properties:
        api_key:
          type: string
      required: [api_key]
    LoginResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        auth_enabled:
          type: boolean
        principal:
          type: string
        auth_mode:
          type: string
    SessionInfo:
      type: object
      properties:
        authenticated:
          type: boolean
        auth_enabled:
          type: boolean
        principal:
          type: string
        auth_mode:
          type: string
...